requirements.txt
pymodbus==3.6.2
python-dotenv==1.0.0
loguru==0.7.2 